Created stopData.csv to store each stop and its list of associated route tags. Also updated getters in Location to return the proper type and the Route constructor now stores the parameter vehicles without making a copy of it (riskier but more efficient).

+++ src/entity/Location.java
@@ -9,11 +9,11 @@ public class Location {
         this.longitude = longitude;
     }
 
-    public double getLatitude() {
+    public float getLatitude() {
         return latitude;
     }
 
-    public double getLongitude() {
+    public float getLongitude() {
         return longitude;
     }
